+++ tools/bun/tests/download_test.rs
@@ -82,6 +82,100 @@ mod bun_tool {
     }
 
     #[tokio::test(flavor = "multi_thread")]
+    async fn supports_linux_arm64_musl() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox
+            .create_plugin_with_config("bun-test", |config| {
+                config.host_with(|host| {
+                    host.os = HostOS::Linux;
+                    host.arch = HostArch::Arm64;
+                    host.libc = HostLibc::Musl;
+                });
+            })
+            .await;
+
+        assert_eq!(
+        plugin
+            .download_prebuilt(DownloadPrebuiltInput {
+                context: PluginContext {
+                    version: VersionSpec::parse("1.4.0").unwrap(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            })
+            .await,
+        DownloadPrebuiltOutput {
+            archive_prefix: Some("bun-linux-aarch64-musl".into()),
+            checksum_url: Some(
+                "https://github.com/oven-sh/bun/releases/download/bun-v1.4.0/SHASUMS256.txt".into()
+            ),
+            download_name: Some("bun-linux-aarch64-musl.zip".into()),
+            download_url:
+                "https://github.com/oven-sh/bun/releases/download/bun-v1.4.0/bun-linux-aarch64-musl.zip"
+                    .into(),
+            ..Default::default()
+        }
+    );
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn supports_linux_x64_musl() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox
+            .create_plugin_with_config("bun-test", |config| {
+                config.host_with(|host| {
+                    host.os = HostOS::Linux;
+                    host.arch = HostArch::X64;
+                    host.libc = HostLibc::Musl;
+                });
+            })
+            .await;
+
+        let result = plugin
+            .download_prebuilt(DownloadPrebuiltInput {
+                context: PluginContext {
+                    version: VersionSpec::parse("1.4.0").unwrap(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            })
+            .await;
+
+        // The baseline suffix is different between boxes in CI!
+        assert!(
+            result.download_url
+                == "https://github.com/oven-sh/bun/releases/download/bun-v1.4.0/bun-linux-x64-musl.zip"
+                || result.download_url
+                    == "https://github.com/oven-sh/bun/releases/download/bun-v1.4.0/bun-linux-x64-musl-baseline.zip"
+        );
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    #[should_panic(expected = "musl is only supported for Bun v1.1.35 and above.")]
+    async fn doesnt_support_musl_before_v1135() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox
+            .create_plugin_with_config("bun-test", |config| {
+                config.host_with(|host| {
+                    host.os = HostOS::Linux;
+                    host.arch = HostArch::Arm64;
+                    host.libc = HostLibc::Musl;
+                });
+            })
+            .await;
+
+        plugin
+            .download_prebuilt(DownloadPrebuiltInput {
+                context: PluginContext {
+                    version: VersionSpec::parse("1.1.34").unwrap(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            })
+            .await;
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
     async fn supports_macos_arm64() {
         let sandbox = create_empty_proto_sandbox();
         let plugin = sandbox
@@ -179,6 +273,81 @@ mod bun_tool {
             ..Default::default()
         }
     );
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn supports_windows_arm64() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox
+            .create_plugin_with_config("bun-test", |config| {
+                config.host(HostOS::Windows, HostArch::Arm64);
+            })
+            .await;
+
+        assert_eq!(
+        plugin
+            .download_prebuilt(DownloadPrebuiltInput {
+                context: PluginContext {
+                    version: VersionSpec::parse("1.4.0").unwrap(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            })
+            .await,
+        DownloadPrebuiltOutput {
+            archive_prefix: Some("bun-windows-aarch64".into()),
+            checksum_url: Some(
+                "https://github.com/oven-sh/bun/releases/download/bun-v1.4.0/SHASUMS256.txt".into()
+            ),
+            download_name: Some("bun-windows-aarch64.zip".into()),
+            download_url:
+                "https://github.com/oven-sh/bun/releases/download/bun-v1.4.0/bun-windows-aarch64.zip"
+                    .into(),
+            ..Default::default()
+        }
+    );
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    #[should_panic(expected = "unsupported architecture arm64 for windows.")]
+    async fn doesnt_support_windows_arm64_before_v1310() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox
+            .create_plugin_with_config("bun-test", |config| {
+                config.host(HostOS::Windows, HostArch::Arm64);
+            })
+            .await;
+
+        plugin
+            .download_prebuilt(DownloadPrebuiltInput {
+                context: PluginContext {
+                    version: VersionSpec::parse("1.3.9").unwrap(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            })
+            .await;
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    #[should_panic(expected = "unsupported OS windows.")]
+    async fn doesnt_support_windows_before_v11() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox
+            .create_plugin_with_config("bun-test", |config| {
+                config.host(HostOS::Windows, HostArch::X64);
+            })
+            .await;
+
+        plugin
+            .download_prebuilt(DownloadPrebuiltInput {
+                context: PluginContext {
+                    version: VersionSpec::parse("1.0.30").unwrap(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            })
+            .await;
     }
 
     #[tokio::test(flavor = "multi_thread")]